editor/libeditor/tests/test_pasteImgTextarea.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>
<window xmlns:html="http://www.w3.org/1999/xhtml"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <body xmlns="http://www.w3.org/1999/xhtml">
    <html:img id="i" src="green.png" />
    <html:textarea id="t"></html:textarea>
  </body>
  <script type="text/javascript"><![CDATA[
    let loaded = new Promise(resolve => addLoadEvent(resolve));
    add_task(async function() {
        await loaded;
        SpecialPowers.setCommandNode(window, document.getElementById("i"));
        SpecialPowers.doCommand(window, "cmd_copyImageContents");
        let input = document.getElementById("t");
        input.focus();
        var controller =
          SpecialPowers.wrap(input).controllers.getControllerForCommand("cmd_paste");
        is(controller.isCommandEnabled("cmd_paste"), false,
           "paste should not be enabled in xul textareas when an image is on the clipboard");
    });
  ]]></script>
</window>
